+++ sw/airborne/firmwares/rotorcraft/Stabilization.java
@@ -1,14 +1,14 @@
 package sw.airborne.firmwares.rotorcraft;
 import static sw.airborne.firmwares.rotorcraft.stabilization.Stabilization_none.*;
 import static sw.airborne.firmwares.rotorcraft.stabilization.Stabilization_rate.*;
-import static sw.airborne.firmwares.rotorcraft.stabilization.Stabilization_attitude_common.*;
+import static sw.airborne.firmwares.rotorcraft.stabilization.Stabilization_attitude_euler_int.*; //TODO Many stabilization_attitude_init()'s
 
 
 public class Stabilization {
 
 	private static final int COMMANDS_NB = 10;
 	private static final boolean STABILIZATION_SKIP_RATE = false;
-	public static int stabilization_cmd[] = new int[COMMANDS_NB];
+	public static long stabilization_cmd[] = new long[COMMANDS_NB];
 
 	public static void stabilization_init() {
 		if(! STABILIZATION_SKIP_RATE){
