+++ sw/airborne/math/Int32Quat.java
@@ -1,8 +1,8 @@
 package sw.airborne.math;
 
 public class Int32Quat {
-	public int qi;
-	public int qx;
-	public int qy;
-	public int qz;
+	public long qi;
+	public long qx;
+	public long qy;
+	public long qz;
 }
