+++ sw/airborne/math/Int32RMat.java
@@ -1,6 +1,6 @@
 package sw.airborne.math;
 
 public class Int32RMat {
-	public int m[] = new int[3*3];
+	public long m[] = new long[3*3];
 
 }
