+++ sw/airborne/math/Int32Rates.java
@@ -1,9 +1,9 @@
 package sw.airborne.math;
 
 public class Int32Rates {
-	public int p; ///< in rad/s^2 with INT32_RATE_FRAC
+	public long p; ///< in rad/s^2 with INT32_RATE_FRAC
 
-	public  int q; ///< in rad/s^2 with INT32_RATE_FRAC
-	public  int r;
+	public  long q; ///< in rad/s^2 with INT32_RATE_FRAC
+	public  long r;
 
 }
